python_new-main/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Core LangGraph/LangChain - Latest compatible versions
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Local LLM Integration
requests>=2.31.0

# Vector Store & Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Memory & Database
redis>=5.0.0
mysql-connector-python>=8.0.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-decouple>=3.8
bcrypt>=4.0.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
aiofiles>=23.0.0
